refactor: move latestBlock to layer2
